requirements.txt
discord.py
autobahn
pyfiglet
wavelink
gcloud-aio-storage
statcord.py-beta
box
psutil
